Added API method to get gateway language
